requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-common==1.1.27
azure-core==1.17.0
azure-functions==1.7.2
azure-nspkg==3.0.2
azure-storage-blob==12.8.1
azure-storage-common==1.4.2
azure-storage-nspkg==3.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
idna==3.2
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.6


